+++ wayland-client/src/proxy.rs
@@ -272,13 +272,13 @@ where
     /// you want to assign this object to its own filter. In which
     /// case you just need to provide the appropriate closure, of
     /// type `FnMut(Main<I>, I::Event)`.
-    pub fn assign_mono<F>(&self, mut f: F)
+    pub fn quick_assign<F>(&self, mut f: F)
     where
         I: Interface + AsRef<Proxy<I>> + From<Proxy<I>> + Sync,
-        F: FnMut(Main<I>, I::Event) + 'static,
+        F: FnMut(Main<I>, I::Event, crate::DispatchData) + 'static,
         I::Event: MessageGroup<Map = crate::ProxyMap>,
     {
-        self.assign(Filter::new(move |(proxy, event), _| f(proxy, event)))
+        self.assign(Filter::new(move |(proxy, event), _, data| f(proxy, event, data)))
     }
 }